제네릭2 패키지 위치 변경
