+++ src/generic/ex3/AnimalHospitalMainV1.java
@@ -1,4 +1,4 @@
-package generic.test.ex3;
+package generic.ex3;
 
 import generic.animal.Cat;
 import generic.animal.Dog;
