+++ src/generic/ex3/AnimalHospitalV2.java
@@ -1,4 +1,4 @@
-package generic.test.ex3;
+package generic.ex3;
 
 public class AnimalHospitalV2<T> {
