+++ src/generic/ex3/CatHospital.java
@@ -1,4 +1,4 @@
-package generic.test.ex3;
+package generic.ex3;
 
 import generic.animal.Cat;
