+++ src/generic/ex3/DogHospital.java
@@ -1,12 +1,11 @@
-package generic.test.ex3;
+package generic.ex3;
 
-import generic.animal.Animal;
+import generic.animal.Dog;
 
-public class AnimalHospitalV1 {
+public class DogHospital {
+    private Dog animal;
 
-    private Animal animal;
-
-    public void set(Animal animal) {
+    public void set(Dog animal) {
         this.animal = animal;
     }
 
@@ -16,7 +15,7 @@ public class AnimalHospitalV1 {
         animal.sound();
     }
 
-    public Animal bigger(Animal target) {
+    public Dog bigger(Dog target) {
         return animal.getSize() > target.getSize() ? animal : target;
     }
 }
