+++ src/generic/ex4/MethodMain3.java
@@ -1,4 +1,4 @@
-package generic.test.ex4;
+package generic.ex4;
 
 import generic.animal.Cat;
 import generic.animal.Dog;
