+++ src/generic/ex5/Box.java
@@ -1,4 +1,4 @@
-package generic.test.ex5;
+package generic.ex5;
 
 public class Box<T> {
